pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-nimc"
version = "0.0.0"
description = "NIMC server creation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hcloud>=2.16.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
web = [
    "pydantic-settings>=2.12.0",
    "bcrypt>=4.0.0,<5.0.0",
    "fastapi>=0.128.0",
    "gunicorn>=25.0.1",
    "uvicorn>=0.40.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "starlette-csrf>=3.0.0",
]
cli = [
    "pydantic-settings>=2.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["nimc/core", "nimc/web", "nimc/cli"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "uvicorn[standard]>=0.40.0",
]
